requirements.txt
# Core
groq
streamlit
python-dotenv
rich

# Data
pandas

# Web
requests
beautifulsoup4

# A2A Protocol
fastapi
uvicorn[standard]
httpx

# Testing
pytest
